requirements.txt
aiofiles==23.2.1
aiogram==2.21
aiohappyeyeballs==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Automat==22.10.0
Babel==2.9.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.8
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
emoji==2.12.1
et-xmlfile==1.1.0
filelock==3.15.4
Flask==2.2.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lxml==5.3.0
magic-filter==1.0.12
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.4.0
plotly==5.23.0
preshed==3.0.9
Protego==0.3.1
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.1
Scrapy==2.11.2
sentencepiece==0.2.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stanza==1.8.2
sympy==1.13.2
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tldextract==5.1.2
toml==0.10.2
torch==2.2.0
tqdm==4.66.5
Twisted==24.7.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==2.2.2
wrapt==1.16.0
xx-sent-ud-sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.7.0/xx_sent_ud_sm-3.7.0-py3-none-any.whl#sha256=aafb609d5a895a62ed9672fbef2aa8061106a4b164a700999a376f8529acc3ad
yarl==1.9.4
zope.interface==7.0.1
